# swagger.yaml
swagger: "2.0"
info:
  description: "This is a sample TODO server."
  version: "1.0.0"
  title: "Swagger TODO"

schemes:
- "http"
paths:
  /todos:
    get:
      summary: "Returns all todos"
      produces:
      - "application/json"
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Todo"
        "400":
          description: "Invalid status value"

  /newTodo:
    post:
      summary: "Creates a new todo item"
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: "string"
      produces:
      - "application/json"
      responses:
        "201":
          description: "created"
          schema:
            $ref: "#/definitions/Todo"
        "400":
          description: "Invalid status value"

  /newSubTodo:
    post:
      summary: "Creates a new subTodo item"
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              parentId:
                type: "string"
                format: "string"
              title:
                type: "string"
      responses:
        "201":
          description: "created"
          schema:
            $ref: "#/definitions/SubTodo"
        "400":
          description: "Invalid status value"
          
  /renameTodo:
    put:
      summary: "Renames a todo item"
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              id:
                type: "string"
                format: "string"
              title:
                type: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Todo"
        "400":
          description: "Invalid status value"
          
  /renameSubTodo:
    put:
      summary: "Renames a subTodo item"
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              id:
                type: "string"
                format: "string"
              title:
                type: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/SubTodo"
        "400":
          description: "Invalid status value"
          
  /tickTodo:
    put:
      summary: "Ticks or unticks a todo item"
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              id:
                type: "string"
                format: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Todo"
        "400":
          description: "Invalid status value"
          
  /tickSubTodo:
    put:
      summary: "Ticks or unticks a subTodo item"
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              id:
                type: "string"
                format: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/SubTodo"
        "400":
          description: "Invalid status value"
          
  /moveSubTodo:
    put:
      summary: "Ticks or unticks a subTodo item"
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              id:
                type: "string"
                format: "string"
              parentId:
                type: "string"
                format: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/SubTodo"
        "400":
          description: "Invalid status value"


definitions:
  Todo:
    type: "object"
    properties:
      id:
        type: "string"
        format: "string"
      title:
        type: "string"
        format: "string"
      isDone:
        type: "boolean"
        default: false
      createdAt:
        type: "string"
        format: "date-time"
      updatedAt:
        type: "string"
        format: "date-time"
      subTodos:
        type: "array"
        items: 
          $ref: '#/definitions/SubTodo'

  SubTodo:
    type: "object"
    properties:
      id:
        type: "string"
        format: "string"
      title:
        type: "string"
        format: "string"
      isDone:
        type: "boolean"
        default: false